# docs/swagger.yaml
openapi: 3.0.4
info:
  title: E-commerce API
  description: |
    Complete e-commerce API with user authentication, product management, 
    shopping cart, orders, payments, and admin functionality.
    
    ## Authentication
    - JWT Bearer token authentication
    - Email verification required for full access
    - Google OAuth support
    
    ## Roles
    - **user**: Regular customer permissions
    - **admin**: Full administrative access
    
    ## Rate Limiting
    - 1000 requests per hour per IP
    - 100 requests per minute for authenticated users
  version: 1.0.0
  termsOfService: https://example.com/terms
  contact:
    name: Abdulrahman Hegazy
    email: a.0hegazy@proton.me
...
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://ecommerce-express-boilerplate.onrender.com/api/v1
    description: Production server

# Global security - applies to all endpoints
security:
  - bearerAuth: []
...
paths:
...
  /reviews:
    post:
      tags:
        - Reviews
      summary: Create a review
      description: Create a new review for a product (only if user has purchased the product)
      operationId: createReview
      requestBody:
        required: true
        description: Review details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
            examples:
              success:
                summary: Create review
                value:
                  productId: "64c9b8d3a7e9a2d8f0c1b236"
                  rating: 5
                  comment: "Excellent product! Great quality and fast delivery."
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Review created
                  value:
                    status: success
                    data:
                      review:
                        id: "64c9b8d3a7e9a2d8f0c1b239"
                        userId: "64c9b8d3a7e9a2d8f0c1b235"
                        productId: "64c9b8d3a7e9a2d8f0c1b236"
                        rating: 5
                        comment: "Excellent product! Great quality and fast delivery."
                        isActive: true
                        createdAt: "2025-10-15T15:00:00.000Z"
                        updatedAt: "2025-10-15T15:00:00.000Z"
                      message: "Review created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - User hasn't purchased the product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                notPurchased:
                  summary: Not purchased
                  value:
                    status: fail
                    data:
                      message: "You can only review products you have purchased"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /reviews/my-reviews:
    get:
      tags:
        - Reviews
      summary: Get user's reviews
      description: Retrieve paginated list of authenticated user's reviews
      operationId: getUserReviews
      parameters:
        - $ref: '#/components/parameters/PaginationPage'
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Number of reviews per page
      responses:
        '200':
          description: User reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: User reviews
                  value:
                    status: success
                    data:
                      reviews:
                        - id: "64c9b8d3a7e9a2d8f0c1b239"
                          userId: "64c9b8d3a7e9a2d8f0c1b235"
                          productId: "64c9b8d3a7e9a2d8f0c1b236"
                          rating: 5
                          comment: "Excellent product! Great quality and fast delivery."
                          isActive: true
                          product:
                            id: "64c9b8d3a7e9a2d8f0c1b236"
                            name: "Wireless Headphones"
                            image: "https://example.com/headphones.jpg"
                            price: 29.99
                          createdAt: "2025-10-15T15:00:00.000Z"
                          updatedAt: "2025-10-15T15:00:00.000Z"
                        - id: "64c9b8d3a7e9a2d8f0c1b240"
                          userId: "64c9b8d3a7e9a2d8f0c1b235"
                          productId: "64c9b8d3a7e9a2d8f0c1b237"
                          rating: 4
                          comment: "Good product but delivery was a bit slow."
                          isActive: true
                          product:
                            id: "64c9b8d3a7e9a2d8f0c1b237"
                            name: "Phone Case"
                            image: "https://example.com/case.jpg"
                            price: 15.99
                          createdAt: "2025-10-14T10:00:00.000Z"
                          updatedAt: "2025-10-14T10:00:00.000Z"
                      pagination:
                        total: 8
                        currentPage: 1
                        totalPages: 1
                        pageSize: 10
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /reviews/product/{productId}:
    get:
      tags:
        - Reviews
      summary: Get product reviews
      description: Retrieve paginated reviews for a specific product
      operationId: getProductReviews
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/PaginationPage'
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Number of reviews per page
      responses:
        '200':
          description: Product reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Product reviews
                  value:
                    status: success
                    data:
                      reviews:
                        - id: "64c9b8d3a7e9a2d8f0c1b239"
                          userId: "64c9b8d3a7e9a2d8f0c1b235"
                          productId: "64c9b8d3a7e9a2d8f0c1b236"
                          rating: 5
                          comment: "Excellent product! Great quality and fast delivery. Would definitely recommend to others."
                          isActive: true
                          user:
                            id: "64c9b8d3a7e9a2d8f0c1b235"
                            name: "John Doe"
                            avatar: "https://via.placeholder.com/150/cccccc/969696?text=User"
                          createdAt: "2025-10-15T15:00:00.000Z"
                          updatedAt: "2025-10-15T15:00:00.000Z"
                        - id: "64c9b8d3a7e9a2d8f0c1b241"
                          userId: "64c9b8d3a7e9a2d8f0c1b242"
                          productId: "64c9b8d3a7e9a2d8f0c1b236"
                          rating: 4
                          comment: "Good sound quality but battery life could be better. Overall satisfied with the purchase."
                          isActive: true
                          user:
                            id: "64c9b8d3a7e9a2d8f0c1b242"
                            name: "Jane Smith"
                            avatar: "https://example.com/avatar2.jpg"
                          createdAt: "2025-10-14T12:00:00.000Z"
                          updatedAt: "2025-10-14T12:00:00.000Z"
                      pagination:
                        total: 15
                        currentPage: 1
                        totalPages: 2
                        pageSize: 10
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /reviews/{id}:
    patch:
      tags:
        - Reviews
      summary: Update review
      description: Update an existing review (only by the review author)
      operationId: updateReview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: objectid
          description: Review ID
          example: "64c9b8d3a7e9a2d8f0c1b239"
      requestBody:
        required: true
        description: Updated review details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
            examples:
              success:
                summary: Update review
                value:
                  rating: 4
                  comment: "Updated review: Good product but could improve battery life."
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Review updated
                  value:
                    status: success
                    data:
                      review:
                        id: "64c9b8d3a7e9a2d8f0c1b239"
                        userId: "64c9b8d3a7e9a2d8f0c1b235"
                        productId: "64c9b8d3a7e9a2d8f0c1b236"
                        rating: 4
                        comment: "Updated review: Good product but could improve battery life."
                        isActive: true
                        createdAt: "2025-10-15T15:00:00.000Z"
                        updatedAt: "2025-10-15T15:30:00.000Z"
                      message: "Review updated successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Reviews
      summary: Delete review
      description: Soft delete a review (only by the review author)
      operationId: deleteReview
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: objectid
          description: Review ID
          example: "64c9b8d3a7e9a2d8f0c1b239"
      responses:
        '200':
          description: Review deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Review deleted
                  value:
                    status: success
                    data:
                      message: "Review deleted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /webhook:
    post:
      tags:
        - Webhooks
      summary: Stripe webhook handler
      description: Handle Stripe webhook events for payment processing and order updates
      operationId: handleWebhook
      security: []  # No auth required - uses Stripe signature
      requestBody:
        required: true
        description: Stripe webhook event payload
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe event data
            examples:
              checkoutCompleted:
                summary: Checkout session completed
                value:
                  id: "evt_123456789"
                  object: "event"
                  api_version: "2023-10-16"
                  created: 1699999999,
                  data: 
                    object:
                      id: "cs_test_123456789"
                      object: "checkout.session"
                      metadata: 
                        userId: "64c9b8d3a7e9a2d8f0c1b235"
                        orderId: "64c9b8d3a7e9a2d8f0c1b248"
                      payment_intent: "pi_123456789"
                      status: "complete"
                  type: "checkout.session.completed"
      parameters:
        - in: header
          name: stripe-signature
          required: true
          schema:
            type: string
          description: Stripe signature for webhook verification
          example: "t=1699999999,v1=abc123def456..."
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
              examples:
                success:
                  summary: Webhook processed
                  value:
                    status: "success"
                    received: true
        '400':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookErrorResponse'
              examples:
                invalidSignature:
                  summary: Invalid signature
                  value:
                    status: "fail"
                    message: "Webhook Error: Invalid signature"
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. For admin endpoints, user must have 'admin' role.
  parameters:
    PaginationPage:
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
...
    ProductId:
      in: path
      name: productId
      required: true
      schema:
        type: string
        format: objectid
      description: Product ID
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Unauthorized. Please log in."
...
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Invalid input data"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Resource not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data: "Internal Server Error"
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, fail, error]
          example: success
        data:
          oneOf:
            - type: object
              description: Contains response data in success and fail cases
              properties:
                message:
                  type: string
                  example: "Successfully completed action"
                accessToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5..."
                user:
                  $ref: '#/components/schemas/User'
            - type: string
              description: Contains error message or additional info in error cases
              example: "Internal server error"
            - type: object
              nullable: true
              description: If no data is available, null can be provided
              example: null
      required:
        - status
        - data
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        data:
          type: string
          example: "Internal Server Error"
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "64c9b8d3a7e9a2d8f0c1b234"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        avatar:
          type: string
          format: uri
          example: "https://via.placeholder.com/150/cccccc/969696?text=User"
        phoneNumber:
          type: string
          example: "+1234567890"
        emailVerified:
          type: boolean
          example: true
        isActive:
          type: boolean
          example: true
        address:
          $ref: '#/components/schemas/Address'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
          example: "2025-10-15T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-10-15T12:30:00.000Z"
    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          default: "US"
          example: "US"
    UserPreferences:
      type: object
      properties:
        newsletter:
          type: boolean
          default: true
          example: true
        notifications:
          type: boolean
          default: true
          example: false
...
    CreateReviewRequest:
      type: object
      required:
        - productId
        - rating
        - comment
      properties:
        productId:
          type: string
          format: objectid
          example: "64c9b8d3a7e9a2d8f0c1b236"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          minLength: 10
          maxLength: 500
          example: "Excellent product! Great quality and fast delivery."
    UpdateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: string
          minLength: 10
          maxLength: 500
          example: "Updated review: Good product but could improve battery life."
...
    WebhookResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, fail]
          example: success
        received:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
    WebhookErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: fail
        message:
          type: string
          example: "Webhook Error: Invalid signature"
...